requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.1.0
Flask-JWT-Extended==4.6.0
Flask-Smorest==0.42.0
Flask-Migrate==4.0.5
Flask-Sock==0.7.0
marshmallow==3.21.1
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.1
redis==5.0.3
psycopg2-binary==2.9.9
passlib==1.7.4
pytest==7.4.4
pytest-flask==1.3.0
Faker==30.3.0
factory_boy==3.3.0
APScheduler==3.10.4
websocket-client==1.8.0 
simple-websocket==1.0.0
websockets==12.0
pytest-cov
coverage-badge

